+++ src/unexploitable_search/tasks/apps/dataset.py
@@ -1,0 +1,75 @@
+import json
+from typing import Any, Dict
+
+import datasets
+from inspect_ai.dataset import MemoryDataset, Sample
+
+
+def record_to_sample(record: Dict[str, Any]) -> Sample:
+    """
+    Convert a record from the apps dataset to a Sample.
+    """
+    # Parse the input_output field
+    try:
+        input_output = (
+            json.loads(record["input_output"]) if record["input_output"] else {}
+        )
+    except (json.JSONDecodeError, TypeError):
+        input_output = {}
+
+    # Get inputs and expected outputs
+    inputs = input_output.get("inputs", [])
+    expected_outputs = input_output.get("outputs", [])
+    fn_name = input_output.get("fn_name", None)
+
+    # Create the sample with the question as input
+    question = record["question"]
+
+    # Add starter code if available
+    starter_code = record.get("starter_code", "")
+    if starter_code:
+        question += f"\n\nStarter code:\n```python\n{starter_code}\n```"
+
+    # Add format instructions
+    if fn_name:
+        question += f"\n\nImplement the function `{fn_name}` that solves this problem. Use standard input/output format."
+    else:
+        question += "\n\nWrite a complete Python program that solves this problem using standard input/output."
+
+    # Get the target - ensure it's always a string
+    target = ""
+    if expected_outputs and len(expected_outputs) > 0:
+        target = str(expected_outputs[0])
+
+    return Sample(
+        input=question,
+        target=target,
+        metadata={
+            "problem_id": record.get("problem_id"),
+            "difficulty": record.get("difficulty"),
+            "inputs": inputs,
+            "expected_outputs": expected_outputs,
+            "fn_name": fn_name,
+            "starter_code": starter_code,
+        },
+    )
+
+
+def load_dataset():
+    # Load raw dataset from PR branch with Parquet conversion
+    raw_dataset = datasets.load_dataset(
+        "codeparrot/apps",
+        revision="refs/pr/4",  # Use PR branch with Parquet conversion
+        split="test",
+    )
+
+    # Transform to Inspect samples manually
+    samples = []
+    for record in raw_dataset:
+        sample = record_to_sample(record)  # type: ignore
+        samples.append(sample)
+
+    # Create memory dataset from transformed samples
+    dataset = MemoryDataset(samples)
+
+    return dataset
