+++ src/unexploitable_search/tasks/apps/scorer.py
@@ -1,0 +1,113 @@
+from inspect_ai.scorer import Score, Target, accuracy, scorer
+from inspect_ai.solver import TaskState
+from inspect_ai.util import sandbox
+
+
+@scorer(metrics=[accuracy()])
+def apps_execution_scorer():
+    """
+    Custom scorer that executes generated Python code and compares output with expected results.
+    """
+
+    async def score(state: TaskState, target: Target) -> Score:
+        # Get the generated code from the model's response
+        generated_code = state.output.completion
+
+        # Get test cases from sample metadata
+        inputs = state.metadata.get("inputs", [])
+        expected_outputs = state.metadata.get("expected_outputs", [])
+
+        if not inputs or not expected_outputs:
+            # No test cases available
+            return Score(
+                value=0.0, explanation="No test cases available for evaluation"
+            )
+
+        # Track results for all test cases
+        passed_tests = 0
+        total_tests = min(len(inputs), len(expected_outputs))
+        execution_details = []
+
+        for i in range(total_tests):
+            test_input = inputs[i]
+            expected_output = expected_outputs[i]
+
+            # Convert input list to string format (each input on a new line)
+            if isinstance(test_input, list):
+                input_str = "\n".join(str(inp) for inp in test_input)
+            else:
+                input_str = str(test_input)
+
+            # Execute the code
+            stdout, stderr = await execute_python_code(generated_code, input_str)
+
+            # Compare output (strip whitespace for comparison)
+            stdout_stripped = stdout.strip()
+            expected_stripped = str(expected_output).strip()
+
+            test_passed = stdout_stripped == expected_stripped
+            if test_passed:
+                passed_tests += 1
+
+            execution_details.append(
+                {
+                    "test_case": i + 1,
+                    "input": input_str,
+                    "expected": expected_stripped,
+                    "actual": stdout_stripped,
+                    "stderr": stderr,
+                    "passed": test_passed,
+                }
+            )
+
+        # Calculate accuracy
+        accuracy = passed_tests / total_tests if total_tests > 0 else 0.0
+
+        # Create explanation
+        explanation = f"Passed {passed_tests}/{total_tests} test cases. "
+        if accuracy < 1.0:
+            failed_cases = [d for d in execution_details if not d["passed"]]
+            if failed_cases:
+                first_failure = failed_cases[0]
+                explanation += f"First failure - Expected: {repr(first_failure['expected'])}, Got: {repr(first_failure['actual'])}"
+                if first_failure["stderr"]:
+                    explanation += f", Error: {first_failure['stderr'][:200]}"
+
+        return Score(
+            value=accuracy,
+            explanation=explanation,
+            metadata={"execution_details": execution_details},
+        )
+
+    return score
+
+
+async def execute_python_code(code: str, input_data: str) -> tuple[str, str]:
+    """
+    Execute Python code with given input in sandbox environment and return (stdout, stderr).
+
+    Args:
+        code: Python code to execute
+        input_data: Input to provide via stdin
+
+    Returns:
+        Tuple of (stdout, stderr) as strings
+    """
+    try:
+        # Write the Python code to a temporary file in the sandbox
+        code_file = "/tmp/generated_code.py"
+        await sandbox().write_file(code_file, code)
+
+        # Execute the Python code with input provided via stdin
+        result = await sandbox().exec(
+            ["python3", code_file],
+            input=input_data,
+            timeout=30,  # 30 second timeout
+        )
+
+        return result.stdout, result.stderr
+
+    except Exception as e:
+        # Capture any execution errors
+        error_msg = f"{type(e).__name__}: {str(e)}"
+        return "", error_msg
