+++ src/unexploitable_search/tasks/apps/tasks.py
@@ -1,0 +1,25 @@
+from inspect_ai import Task, task
+from inspect_ai.solver import generate, system_message
+
+from unexploitable_search.tasks.apps.dataset import load_dataset
+from unexploitable_search.tasks.apps.prompts import VANILLA_SYSTEM_PROMPT
+from unexploitable_search.tasks.apps.scorer import apps_execution_scorer
+
+
+@task
+def problem_solver():
+    """
+    APPS dataset task that generates and executes Python code to solve programming problems.
+    """
+    dataset = load_dataset()
+
+    # System message to guide the model
+    return Task(
+        dataset=dataset,
+        solver=[
+            system_message(VANILLA_SYSTEM_PROMPT),
+            generate(),  # LLM just generates code, no execution during generation
+        ],
+        scorer=apps_execution_scorer(),  # Use our custom execution scorer
+        sandbox="docker",
+    )
