--- src/unexploitable_search/tasks/apps_task.py
@@ -1,219 +1,0 @@
-import json
-from typing import Any, Dict
-
-import datasets
-from inspect_ai import Task, task
-from inspect_ai.dataset import MemoryDataset, Sample
-from inspect_ai.scorer import Score, Target, accuracy, scorer
-from inspect_ai.solver import TaskState, generate, system_message
-from inspect_ai.util import sandbox
-
-# Desired pipeline
-# 1. Write code that would solve the problem given the question
-# 2. Run the code within a sandbox environment, passing the input of the problem
-# 3. Compare the output of the code with the target output
-
-
-def record_to_sample(record: Dict[str, Any]) -> Sample:
-    """
-    Convert a record from the apps dataset to a Sample.
-    """
-    # Parse the input_output field
-    try:
-        input_output = (
-            json.loads(record["input_output"]) if record["input_output"] else {}
-        )
-    except (json.JSONDecodeError, TypeError):
-        input_output = {}
-
-    # Get inputs and expected outputs
-    inputs = input_output.get("inputs", [])
-    expected_outputs = input_output.get("outputs", [])
-    fn_name = input_output.get("fn_name", None)
-
-    # Create the sample with the question as input
-    question = record["question"]
-
-    # Add starter code if available
-    starter_code = record.get("starter_code", "")
-    if starter_code:
-        question += f"\n\nStarter code:\n```python\n{starter_code}\n```"
-
-    # Add format instructions
-    if fn_name:
-        question += f"\n\nImplement the function `{fn_name}` that solves this problem. Use standard input/output format."
-    else:
-        question += "\n\nWrite a complete Python program that solves this problem using standard input/output."
-
-    # Get the target - ensure it's always a string
-    target = ""
-    if expected_outputs and len(expected_outputs) > 0:
-        target = str(expected_outputs[0])
-
-    return Sample(
-        input=question,
-        target=target,
-        metadata={
-            "problem_id": record.get("problem_id"),
-            "difficulty": record.get("difficulty"),
-            "inputs": inputs,
-            "expected_outputs": expected_outputs,
-            "fn_name": fn_name,
-            "starter_code": starter_code,
-        },
-    )
-
-
-async def execute_python_code(code: str, input_data: str) -> tuple[str, str]:
-    """
-    Execute Python code with given input in sandbox environment and return (stdout, stderr).
-
-    Args:
-        code: Python code to execute
-        input_data: Input to provide via stdin
-
-    Returns:
-        Tuple of (stdout, stderr) as strings
-    """
-    try:
-        # Write the Python code to a temporary file in the sandbox
-        code_file = "/tmp/generated_code.py"
-        await sandbox().write_file(code_file, code)
-
-        # Execute the Python code with input provided via stdin
-        result = await sandbox().exec(
-            ["python3", code_file],
-            input=input_data,
-            timeout=30,  # 30 second timeout
-        )
-
-        return result.stdout, result.stderr
-
-    except Exception as e:
-        # Capture any execution errors
-        error_msg = f"{type(e).__name__}: {str(e)}"
-        return "", error_msg
-
-
-@scorer(metrics=[accuracy()])
-def apps_execution_scorer():
-    """
-    Custom scorer that executes generated Python code and compares output with expected results.
-    """
-
-    async def score(state: TaskState, target: Target) -> Score:
-        # Get the generated code from the model's response
-        generated_code = state.output.completion
-
-        # Get test cases from sample metadata
-        inputs = state.metadata.get("inputs", [])
-        expected_outputs = state.metadata.get("expected_outputs", [])
-
-        if not inputs or not expected_outputs:
-            # No test cases available
-            return Score(
-                value=0.0, explanation="No test cases available for evaluation"
-            )
-
-        # Track results for all test cases
-        passed_tests = 0
-        total_tests = min(len(inputs), len(expected_outputs))
-        execution_details = []
-
-        for i in range(total_tests):
-            test_input = inputs[i]
-            expected_output = expected_outputs[i]
-
-            # Convert input list to string format (each input on a new line)
-            if isinstance(test_input, list):
-                input_str = "\n".join(str(inp) for inp in test_input)
-            else:
-                input_str = str(test_input)
-
-            # Execute the code
-            stdout, stderr = await execute_python_code(generated_code, input_str)
-
-            # Compare output (strip whitespace for comparison)
-            stdout_stripped = stdout.strip()
-            expected_stripped = str(expected_output).strip()
-
-            test_passed = stdout_stripped == expected_stripped
-            if test_passed:
-                passed_tests += 1
-
-            execution_details.append(
-                {
-                    "test_case": i + 1,
-                    "input": input_str,
-                    "expected": expected_stripped,
-                    "actual": stdout_stripped,
-                    "stderr": stderr,
-                    "passed": test_passed,
-                }
-            )
-
-        # Calculate accuracy
-        accuracy = passed_tests / total_tests if total_tests > 0 else 0.0
-
-        # Create explanation
-        explanation = f"Passed {passed_tests}/{total_tests} test cases. "
-        if accuracy < 1.0:
-            failed_cases = [d for d in execution_details if not d["passed"]]
-            if failed_cases:
-                first_failure = failed_cases[0]
-                explanation += f"First failure - Expected: {repr(first_failure['expected'])}, Got: {repr(first_failure['actual'])}"
-                if first_failure["stderr"]:
-                    explanation += f", Error: {first_failure['stderr'][:200]}"
-
-        return Score(
-            value=accuracy,
-            explanation=explanation,
-            metadata={"execution_details": execution_details},
-        )
-
-    return score
-
-
-@task
-def problem_solver():
-    """
-    APPS dataset task that generates and executes Python code to solve programming problems.
-    """
-    # Load raw dataset from PR branch with Parquet conversion
-    raw_dataset = datasets.load_dataset(
-        "codeparrot/apps",
-        revision="refs/pr/4",  # Use PR branch with Parquet conversion
-        split="test",
-    )
-
-    # Transform to Inspect samples manually
-    samples = []
-    for record in raw_dataset:
-        sample = record_to_sample(record)  # type: ignore
-        samples.append(sample)
-
-    # Create memory dataset from transformed samples
-    dataset = MemoryDataset(samples)
-
-    # System message to guide the model
-    system_msg = """You are an expert programmer. Given a programming problem, write a complete Python solution.
-
-Instructions:
-- Read the problem carefully
-- Write clean, correct Python code
-- Use standard input/output (input() and print())
-- Handle edge cases
-- Ensure your code is efficient and readable
-- Only provide the code, no explanations unless asked
-
-Your response should contain only the Python code that solves the problem."""
-
-    return Task(
-        dataset=dataset,
-        solver=[
-            system_message(system_msg),
-            generate(),  # LLM just generates code, no execution during generation
-        ],
-        scorer=apps_execution_scorer(),  # Use our custom execution scorer
-        sandbox="docker",
-    )
